refactor: icones sociais de compartilhamento tem seu proprio componente
